Move vacuum and its files into its own package (integrations.roborock)
