+++ miio/integrations/roborock/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .vacuum import Vacuum, VacuumException, VacuumStatus
